+++ CHANGELOG.md
@@ -5,9 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Symbiosis cards to Inscription
+
 ## [1.1.6] 2024-12-31
 ### Added
-- Ascension cards added to Inscription
+- Ascension cards to Inscription
 
 ## [1.1.5] 2024-12-26
 ### Added